feat: added handler to the deployment
