+++ src/commands/deployment/execute.rs
@@ -138,6 +138,10 @@ struct Commit {
     message_headline: String,
 }
 
+fn capitalize_first_letter(o: &str) -> String {
+    o[0..1].to_uppercase() + &o[1..]
+}
+
 #[wukong_telemetry(command_event = "deployment_execute")]
 pub async fn handle_execute(
     context: Context,
@@ -307,6 +311,32 @@ pub async fn handle_execute(
         selected_version = version_selections[selected_version_index].to_string();
 
         println!("You've selected `{selected_version}` as the deployment version.\n");
+    }
+
+    let deployment_status = get_deployment_status(
+        &mut client,
+        &current_application,
+        &selected_namespace,
+        &selected_version,
+    )
+    .await?;
+
+    println!("Deployment status: {}\n", deployment_status);
+
+    if deployment_status != "SUCCEEDED" {
+        let agree_to_continue = Confirm::with_theme(&ColorfulTheme::default())
+                .with_prompt(
+                    format!(
+                    "It seems the {0} deployment is not in a stable state, are you still want to proceed with the {0} deployment ?",
+                        capitalize_first_letter(&selected_version).green(),
+                    )
+                )
+                .default(false)
+                .interact()?;
+
+        if !agree_to_continue {
+            return Ok(false);
+        }
     }
 
     if let Some(artifact) = artifact {
@@ -627,4 +657,34 @@ fn generate_two_columns_build_selection(cd_pipeline: &CdPipelineWithBuilds) -> V
         .iter_mut()
         .for_each(|each| each.left_width = width);
     two_columns
+}
+
+async fn get_deployment_status(
+    client: &mut QueryClient,
+    application: &str,
+    namespace: &str,
+    version: &str,
+) -> Result<String, CliError> {
+    let deployments = client
+        .fetch_cd_pipeline(application, namespace, version)
+        .await?
+        .data
+        .unwrap()
+        .cd_pipeline;
+
+    let latest_deployment = deployments
+        .iter()
+        .find(|deployment| deployment.last_successfully_deployed_artifact.is_some());
+
+    if latest_deployment.is_none() {
+        Ok(String::from("TERMINAL"))
+    } else {
+        let deployment_status = latest_deployment.unwrap().status.clone();
+
+        if deployment_status.is_none() {
+            Ok(String::from("TERMINAL"))
+        } else {
+            Ok(deployment_status.unwrap())
+        }
+    }
 }
